move StoreManager to utils package
